+++ deployment_clusters/lib/stacks/cluster-stack.ts
@@ -1,11 +1,7 @@
 import { Stack, StackProps, aws_eks as eks, aws_ec2 as ec2} from 'aws-cdk-lib';
 import { Construct } from 'constructs';
-import * as cdk from 'aws-cdk-lib';
-import * as iam from 'aws-cdk-lib/aws-iam';
 import { Vpc } from 'aws-cdk-lib/aws-ec2';
-import { Version } from 'aws-cdk-lib/aws-lambda';
-import { AwsAuth, KubernetesVersion, Nodegroup } from 'aws-cdk-lib/aws-eks';
-import { CpuArchitecture } from 'aws-cdk-lib/aws-ecs';
+import { AwsAuth, KubernetesVersion } from 'aws-cdk-lib/aws-eks';
 import { Role } from 'aws-cdk-lib/aws-iam';
 
 
@@ -49,7 +45,6 @@ export class ClusterStack extends Stack {
       this.cluster = new eks.FargateCluster(this, props.name + '-Cluster', {
         clusterName: props.name,
         vpc: props.vpc,
-        // defaultCapacity: 0,  // we want to manage capacity our selves
         version: props.version,
         clusterLogging: logging
       });
