+++ deployment_clusters/lib/utils/vpc-stack.ts
@@ -1,8 +1,7 @@
 import { Stack, StackProps, aws_eks as eks, aws_ec2 as ec2} from 'aws-cdk-lib';
 import * as iam from 'aws-cdk-lib/aws-iam'
 import { Construct } from 'constructs';
-import { readFileSync, writeFileSync } from 'fs';
-import { Effect, ManagedPolicy, PolicyStatement } from 'aws-cdk-lib/aws-iam';
+import { ManagedPolicy } from 'aws-cdk-lib/aws-iam';
 // const yaml = require('js-yaml')
